connectivity: always check for log errors
Currently, we check for errors in Cilium agent logs only when unsafe
tests are enabled. Let's promote this check to be always performed, to
ensure that we catch unexpected errors early (as unsafe tests are only
enabled in a limited set of CI workflows) given that it does not perform
any unsafe operation.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>